split alter
